sm4: 2018 edition and `block-cipher` crate upgrade
Upgrades to Rust 2018 edition and the (now 2018 edition) `block-cipher`
crate (formerly `block-cipher-trait`).
