+++ sm4/Cargo.toml
@@ -7,16 +7,14 @@ description = "SM4 block cipher algorithm"
 documentation = "https://docs.rs/sm4"
 repository = "https://github.com/RustCrypto/block-ciphers/tree/master/rc2"
 readme = "README.md"
+edition = "2018"
 keywords = ["crypto", "sm4", "block-cipher"]
 categories = ["cryptography"]
 
 [dependencies]
-block-cipher-trait = "0.6"
+block-cipher = "= 0.7.0-pre"
 byteorder = { version = "1", default-features = false }
 opaque-debug = "0.2"
 
 [dev-dependencies]
-block-cipher-trait = { version = "0.6", features = ["dev"] }
-
-[badges]
-maintenance = { status = "experimental" }
+block-cipher = { version = "= 0.7.0-pre", features = ["dev"] }
