+++ sm4/benches/lib.rs
@@ -1,7 +1,7 @@
 #![no_std]
 #![feature(test)]
 #[macro_use]
-extern crate block_cipher_trait;
+extern crate block_cipher;
 extern crate sm4;
 
 bench!(sm4::Sm4, 16);
